Usersettings with default values in preprocessor
Remove usersettings from MIDDLEWARE and CONTEXT_PREPROCESSOR and
create a function (called get_site_setting) and a class (called
DefaultSiteSettings) to replace it. The function is used to
provide usersettings defaults in python code. The class is used
in blognajd context_processors to inject usersettings with
default values in templates.