Fix template tag, adds a simple migration
Fix taggit_templatetags2 tag 'get_tags_for_object'.
The upstream tag doesn't take into account the
content_type of the object. So when several models
are being tagged, it retrieves tags for all the
objects with the same ID, no matter the content_type.
The fix consist of replicating the functionality
in the blognajd_tags module adding the content_type.

The migration provides the model blognajd.sitesettings
with defaults for all the fields, so that the
functionality get_site_setting retrieves default
values when the table for that model is empty.